refactor(trusty-installer): split install.rs report/summary into install_report.rs
CI's 500-line file-size cap flagged install.rs at 504 SLOC after the
required/optional graceful-degrade fix. Extracts the report/outcome
types, the --dry-run preview, and the human-summary renderer
(InstallOutcome, InstallReport, DryRunMember, DryRunReport,
plans_service_bootstrap, plans_mpm_supervisor_bootstrap,
build_dry_run_report, print_dry_run, print_human_summary) into a new
sibling module install_report.rs, mirroring this crate's existing
commands/up/report.rs split for the same command family.

Behavior-preserving only — no logic changes. install.rs keeps run(),
install_all(), install_one(), and the cargo-bin-dir helpers; the new
module's items are pub(super) so install.rs and install_tests.rs (a
descendant of the install module) both see them, matching the existing
cli.rs/cli_tests.rs sibling-file convention already documented in this
file.

scripts/check_line_cap.sh: 0 violations (was 1). cargo test -p
trusty-installer: 431 passed, 0 failed. clippy -D warnings: clean.
fmt --check: clean.
